application/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework

Flask==1.0.2
mysql-connector-python==8.0.16
